+++ man-roxygen/chunk_name_prefix.R
@@ -1,0 +1,1 @@
+#' @param chunk_name_prefix Character string with prefix of chunknames that will be removed. Default: NULL (indicating all chunknames will be removed except the one named `setup`)